% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aipw_balancer}
\alias{aipw_balancer}
\title{Calculate covariate balance metrics}
\usage{
aipw_balancer(XX, W, e.hat)
}
\arguments{
\item{XX}{Matrix of covariates.}

\item{W}{Vector of treatment assignments.}

\item{e.hat}{Vector of estimated propensity scores.}
}
\value{
A list containing adjusted and unadjusted covariate balance metrics.
}
\description{
Calculates unadjusted and adjusted covariate balance metrics using AIPW method.
}
