% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalwiz-package.R
\docType{package}
\name{causalwiz-package}
\alias{causalwiz}
\alias{causalwiz-package}
\title{causalwiz: Tools for Causal Inference Analysis}
\description{
causalwiz is a comprehensive toolkit for causal inference analysis in R.
The package provides functions for estimating treatment effects using various
statistical methods, features built-in plotting functions and utilities designed
for causal inference workflows in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://carminemnc.github.io/causalwiz}
  \item Report bugs at \url{https://github.com/carminemnc/causalwiz/issues}
}

}
\author{
Carmine Minichini \email{carmine.mnc@gmail.com}
}
\keyword{internal}
