% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalwiz-package.R
\docType{package}
\name{causalwiz}
\alias{causalwiz-package}
\alias{causalwiz}
\title{causalwiz: A Package for Causal Inference Analysis}
\description{
The causalwiz package provides tools for causal inference analysis,
implementing various methods including Inverse Probability Weighting (IPW)
and Augmented Inverse Probability Weighting (AIPW). It includes utilities
for estimating treatment effects, analyzing covariate balance, and
visualizing results.
}
\details{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carminemnc/causalwiz}
  \item Report bugs at \url{https://github.com/carminemnc/causalwiz/issues}
}

}
