% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_formula}
\alias{get_formula}
\title{Generate formula for model specification}
\usage{
get_formula(vars, model_spec = "linear")
}
\arguments{
\item{vars}{Character vector of variable names.}

\item{model_spec}{Character string specifying the model type: 'linear', 'interaction', or 'splines'.}
}
\value{
A formula object.
}
\description{
Provides a formula for model specification based on given variables and model type.
}
