% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{welfare_small}
\alias{welfare_small}
\title{Welfare Dataset (Small Version)}
\format{
A data frame with X rows and 9 columns:
\describe{
  \item{y}{Outcome variable}
  \item{w}{Treatment variable}
  \item{age}{Age of the respondent}
  \item{polviews}{Political views of the respondent}
  \item{income}{Income of the respondent}
  \item{educ}{Education level of the respondent}
  \item{marital}{Marital status of the respondent}
  \item{sex}{Sex of the respondent}
}
}
\source{
Smith, Tom W., Michael Davern, Jeremy Freese, and Stephen L. Morgan.
General Social Surveys, 1972-2014.
Principal Investigator, Tom W. Smith; Co-Principal Investigators, Michael Davern,
Jeremy Freese, and Stephen L. Morgan, NORC ed. Chicago: NORC, 2018.
}
\usage{
welfare_small
}
\description{
A subset of the General Social Survey (GSS) data used for causal inference analysis.
The GSS is a high-quality, representative sample of the adult population of the United States,
conducted regularly since 1972 to monitor trends in attitudes, behaviors, and attributes.
}
\examples{
data(welfare_small)
head(welfare_small)
summary(welfare_small)
}
\references{
Smith, T. W. (2016). The General Social Surveys. NORC at the University of Chicago.
GSS Project Report No. 32.
}
\keyword{datasets}
