% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalwiz-package.R
\docType{package}
\name{causalwiz-package}
\alias{causalwiz}
\alias{causalwiz-package}
\title{causalwiz: Tools for Causal Inference Analysis}
\description{
causalwiz provides a comprehensive toolkit for causal inference analysis in R.
The package implements modern methods for estimating treatment effects, including
Inverse Probability Weighting (IPW) and Augmented Inverse Probability Weighting (AIPW),
with built-in diagnostics for covariate balance and propensity score analysis.
}
\details{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
}
\section{Getting Started}{

See \code{vignette("chapter0", package = "causalwiz")} for an introduction.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://carminemnc.github.io/causalwiz}
  \item Report bugs at \url{https://github.com/carminemnc/causalwiz/issues}
}

}
\keyword{internal}
