% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cov_bal_plot}
\alias{cov_bal_plot}
\title{Create covariates balance plot}
\usage{
cov_bal_plot(XX, unadjusted, adjusted)
}
\arguments{
\item{XX}{Matrix of covariates}

\item{unadjusted}{Vector of unadjusted values}

\item{adjusted}{Vector of adjusted values}
}
\value{
A ggplot object
}
\description{
Creates a plot comparing standardized mean differences (SMD) for covariates
before and after adjustment. The plot displays two panels: one showing unadjusted
SMDs and another showing adjusted SMDs. Values closer to zero indicate better
balance between treatment and control groups.
}
